requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
Babel==2.15.0
beautifulsoup4==4.12.2
black==23.7.0
bleach==6.1.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.2.2
contourpy==1.1.0
coverage==7.2.7
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.0
executing==2.0.1
fastapi==0.100.0
fastjsonschema==2.19.1
filelock==3.12.2
flake8==6.0.0
Flask==2.3.2
Flask-RESTful==0.3.10
fonttools==4.41.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gensim==4.3.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
iniconfig==2.0.0
inquirerpy==0.3.4
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.4
lightning-utilities==0.9.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook==7.2.0
notebook_shim==0.2.4
numpy==1.24.2
overrides==7.7.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.11.1
patsy==0.5.3
pfzy==0.3.4
Pillow==10.0.0
platformdirs==3.9.1
pluggy==1.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.46
psutil==5.9.8
psycopg2-binary==2.9.6
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.1.1
pydantic_core==2.4.0
pyflakes==3.0.1
pygame==2.6.1
Pygments==2.18.0
pyparsing==3.0.9
pytest==7.2.1
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytorch-lightning==2.0.6
pytz==2023.3
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
ruff==0.0.280
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
Send2Trash==1.8.3
six==1.16.0
smart-open==7.0.5
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.0
sympy==1.12
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.2.0
tinycss2==1.3.0
tomli==2.0.1
torch==2.0.1
torchmetrics==1.0.1
tornado==6.4
tqdm==4.65.0
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.7.1
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.23.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.3.6
widgetsnbextension==4.0.11
wrapt==1.16.0
yarl==1.9.2
